requirements.txt
PyQt6>=6.5.0
requests>=2.31.0
python-vlc>=3.0.18121
lxml>=4.9.0
python-dateutil>=2.8.2